backend2/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.27
python-telegram-bot==21.9
redis==5.2.1
tonsdk==1.0.15
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.1
loguru==0.7.3
prometheus-client==0.19.0 